compress pred.list.red

+++ cpu_2d/compression/cpusimd.hh
@@ -28,7 +28,7 @@ public:
     CpuSimd();
     void debugCompression(T *fq, const size_t size) const;
     inline void compress(T *fq_64, const size_t size, T_C **compressed_fq_64, size_t &compressedsize) const ;
-    inline void decompress(T_C *compressed_fq_64, const int size,
+    inline void decompress(T_C *compressed_fq_64, const size_t size,
                     /*Out*/ T **uncompressed_fq_64, /*In Out*/size_t &uncompressedsize) const;
     void verifyCompression(const T *fq, const T *uncompressed_fq_64, size_t uncompressedsize) const;
     inline bool isCompressed(const size_t originalsize, const size_t compressedsize) const;
@@ -118,7 +118,7 @@ inline void CpuSimd<T, T_C>::compress(T * restrict fq_64, const size_t size, T_C
 }
 
 template <typename T, typename T_C>
-inline void CpuSimd<T, T_C>::decompress(T_C * restrict compressed_fq_32, const int size,
+inline void CpuSimd<T, T_C>::decompress(T_C * restrict compressed_fq_32, const size_t size,
                                  /*Out*/ T ** restrict uncompressed_fq_64, /*In Out*/size_t &uncompressedsize) const
 {
    if (isCompressed(uncompressedsize, size))
